move AssessmentViewPropsBuilder to a seperate file
